refactor: implement player join and quit messages in ChatModule with customizable formats
